child/requirements.txt
google-cloud-storage
google-cloud-pubsub
bs4
requests
base64